Package less names for view schema types, started cleanup
 * Separate files for the component factories
 * package private and final to show intent
 * noted down some problems
