+++ sdk/spring-sdk/src/main/scala/kalix/springsdk/impl/ViewDescriptorFactory.scala
@@ -1,0 +1,200 @@
+/*
+ * Copyright 2021 Lightbend Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package kalix.springsdk.impl
+
+import kalix.MethodOptions
+import kalix.springsdk.annotations.Query
+import kalix.springsdk.annotations.Subscribe
+import kalix.springsdk.annotations.Table
+import kalix.springsdk.impl.ComponentDescriptorFactory.eventingInForValueEntity
+import kalix.springsdk.impl.ComponentDescriptorFactory.findValueEntityType
+import kalix.springsdk.impl.ComponentDescriptorFactory.hasValueEntitySubscription
+import kalix.springsdk.impl.reflection.KalixMethod
+import kalix.springsdk.impl.reflection.NameGenerator
+import kalix.springsdk.impl.reflection.RestServiceIntrospector
+import kalix.springsdk.impl.reflection.RestServiceMethod
+import kalix.springsdk.impl.reflection.VirtualServiceMethod
+
+import java.lang.reflect.ParameterizedType
+
+private[impl] object ViewDescriptorFactory {
+  private def valueEntityStateClassOf(valueEntityClass: Class[_]): Class[_] = {
+    valueEntityClass.getGenericSuperclass
+      .asInstanceOf[ParameterizedType]
+      .getActualTypeArguments
+      .head
+      .asInstanceOf[Class[_]]
+  }
+}
+
+private[impl] final class ViewDescriptorFactory[T](val component: Class[T]) extends ComponentDescriptorFactory[T] {
+  import ViewDescriptorFactory._
+
+  override def buildDescriptor(nameGenerator: NameGenerator): ComponentDescriptor = {
+    val hasTypeLevelValueEntitySubs = component.getAnnotation(classOf[Subscribe.ValueEntity]) != null
+    val hasMethodLevelValueEntitySubs = component.getMethods.exists(hasValueEntitySubscription)
+    // View class type parameter declares table type
+    val tableType: Class[_] =
+      component.getGenericSuperclass.asInstanceOf[ParameterizedType].getActualTypeArguments.head.asInstanceOf[Class[_]]
+    val tableTypeDescriptor = ProtoMessageDescriptors.generateMessageDescriptors(tableType)
+
+    def addTableOptionsToUpdateMethod(builder: MethodOptions.Builder, transform: Boolean) = {
+      val tableName: String = component.getAnnotation(classOf[Table]).value()
+      if (tableName == null || tableName.trim.isEmpty) {
+        throw new IllegalArgumentException(
+          s"Table name for [${component.getName}] is empty, must be a non-empty string.")
+      }
+
+      val update = kalix.View.Update
+        .newBuilder()
+        .setTable(tableName)
+        .setTransformUpdates(transform)
+
+      val jsonSchema = kalix.JsonSchema
+        .newBuilder()
+        .setOutput(tableTypeDescriptor.mainMessageDescriptor.getName)
+        .build()
+
+      val view = kalix.View
+        .newBuilder()
+        .setUpdate(update)
+        .setJsonSchema(jsonSchema)
+        .build()
+      builder.setView(view)
+    }
+
+    val updateMethods = {
+      if (hasTypeLevelValueEntitySubs && hasMethodLevelValueEntitySubs)
+        throw new IllegalArgumentException(
+          "Mixed usage of @Subscribe.ValueEntity annotations. " +
+          "You should either use it at type level or at method level, not both.")
+
+      if (hasTypeLevelValueEntitySubs) {
+        // create a virtual method
+        val methodOptionsBuilder = kalix.MethodOptions.newBuilder()
+
+        // validate
+        val valueEntityClass: Class[_] =
+          component.getAnnotation(classOf[Subscribe.ValueEntity]).value().asInstanceOf[Class[_]]
+        val entityStateClass = valueEntityStateClassOf(valueEntityClass)
+        if (entityStateClass != tableType)
+          throw new IllegalArgumentException(
+            s"View subscribes to ValueEntity [${valueEntityClass.getName}] and subscribes to state changes " +
+            s"which will be of type [${entityStateClass.getName}] but view type parameter is [${tableType.getName}] which does not match, " +
+            "the types of the entity and the subscribing must be the same.")
+
+        val entityType = findValueEntityType(component)
+        methodOptionsBuilder.setEventing(eventingInForValueEntity(entityType))
+
+        addTableOptionsToUpdateMethod(methodOptionsBuilder, false)
+        val kalixOptions = methodOptionsBuilder.build()
+
+        Seq(
+          KalixMethod(VirtualServiceMethod(component, "OnChange"))
+            .withKalixOptions(kalixOptions))
+
+      } else {
+        var previousValueEntityClass: Option[Class[_]] = None
+
+        component.getMethods
+          .filter(hasValueEntitySubscription)
+          .map { method =>
+            // validate
+            val valueEntityClass = method.getAnnotation(classOf[Subscribe.ValueEntity]).value().asInstanceOf[Class[_]]
+            previousValueEntityClass match {
+              case Some(`valueEntityClass`) => // ok
+              case Some(other) =>
+                throw new IllegalArgumentException(
+                  s"All update methods must return the same type, but [${method.getName}] returns [${valueEntityClass.getName}] while a prevous update method returns [${other.getName}]")
+              case None => previousValueEntityClass = Some(valueEntityClass)
+            }
+            // FIXME validate that transform method accepts value entity state type
+
+            // event sourced or topic subscription updates
+            val methodOptionsBuilder = kalix.MethodOptions.newBuilder()
+            methodOptionsBuilder.setEventing(eventingInForValueEntity(method))
+            addTableOptionsToUpdateMethod(methodOptionsBuilder, true)
+
+            KalixMethod(RestServiceMethod(method)).withKalixOptions(methodOptionsBuilder.build())
+
+          }
+          .toSeq
+      }
+    }
+
+    // we only take methods with Query annotations and Spring REST annotations
+    val (queryMethod, queryResultDescriptor) = {
+      val annotatedMethods = RestServiceIntrospector
+        .inspectService(component)
+        .methods
+        .filter(_.javaMethod.getAnnotation(classOf[Query]) != null)
+      if (annotatedMethods.isEmpty)
+        throw new IllegalArgumentException(
+          "No method annotated with @Query found. " +
+          "Views should have a method annotated with @Query")
+      if (annotatedMethods.size > 1)
+        throw new IllegalArgumentException("Views can have only one method annotated with @Query")
+
+      val annotatedMethod = annotatedMethods.head
+
+      val returnType = annotatedMethod.javaMethod.getReturnType
+      val returnTypeDescriptor =
+        if (returnType == tableType) tableTypeDescriptor
+        else ProtoMessageDescriptors.generateMessageDescriptors(returnType)
+
+      val queryStr = annotatedMethod.javaMethod.getAnnotation(classOf[Query]).value()
+
+      val query = kalix.View.Query
+        .newBuilder()
+        .setQuery(queryStr)
+        .build()
+
+      val jsonSchema = kalix.JsonSchema
+        .newBuilder()
+        .setInput(annotatedMethod.requestProtoMessageName)
+        .setJsonBodyInputField("json_body") // is it safe to always specify?
+        .setOutput(returnTypeDescriptor.mainMessageDescriptor.getName)
+        .build()
+
+      val view = kalix.View
+        .newBuilder()
+        .setJsonSchema(jsonSchema)
+        .setQuery(query)
+        .build()
+
+      val builder = kalix.MethodOptions.newBuilder()
+      builder.setView(view)
+      val methodOptions = builder.build()
+
+      (KalixMethod(annotatedMethod).withKalixOptions(methodOptions), returnTypeDescriptor)
+    }
+
+    val kalixMethods: Seq[KalixMethod] = queryMethod +: updateMethods
+
+    val serviceName = nameGenerator.getName(component.getSimpleName)
+    val componentDescriptor =
+      kalixMethods.foldLeft(new ComponentDescriptor(serviceName, component.getPackageName, nameGenerator)) {
+        (desc, method) =>
+          desc.withMethod(method)
+      }
+    componentDescriptor.withMessageDescriptor(tableTypeDescriptor)
+    if (tableTypeDescriptor != queryResultDescriptor)
+      componentDescriptor.withMessageDescriptor(queryResultDescriptor)
+
+    componentDescriptor
+  }
+}
